+++ src/main/java/com/masters/waterways/daos/shipmodeldao.java
@@ -8,11 +8,11 @@ public interface shipmodeldao {
 	
     int save(ShipModel shipmodel);
 	
-	int update(ShipModel shipmodel,long id);
+	int update(ShipModel shipmodel,int id);
 	
-	int delete(long id);
+	int delete(int id);
 	
 	List<ShipModel> getall();
 	
-	ShipModel getbyid(long id);
+	ShipModel getbyid(int id);
 }
